Process_metier_generation/PROD/requirements.txt
flask
flask-cors
python-dotenv
boto3
botocore
openai